refactor: 신뢰도 타입을 Double에서 Float으로 변경

+++ android/data/src/main/java/com/ddangddangddang/data/model/request/ReviewRequest.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class ReviewRequest(
     val auctionId: Long,
     val targetId: Long,
-    val score: Double,
+    val score: Float,
     val content: String,
 )
